feat: added update button on new version
